+++ ysqlite3/vfs/sqlite3_file_wrapper.hpp
@@ -14,8 +14,8 @@ namespace vfs {
 class sqlite3_file_wrapper : public file
 {
 public:
-	sqlite3_file_wrapper(file_format format, std::shared_ptr<sqlite3_file> parent) noexcept
-	    : file{ format }, _parent{ std::move(parent) }
+	sqlite3_file_wrapper(const char* name, file_format format, std::shared_ptr<sqlite3_file> parent) noexcept
+	    : file{ name, format }, _parent{ std::move(parent) }
 	{}
 	void close() override
 	{
@@ -74,6 +74,30 @@ public:
 	int device_characteristics() const noexcept override
 	{
 		return _parent->pMethods->xDeviceCharacteristics(_parent.get());
+	}
+	void shm_map(int page, int page_size, bool is_write, void volatile** mapped_memory) override
+	{
+		_check_error(_parent->pMethods->xShmMap(_parent.get(), page, page_size, is_write, mapped_memory));
+	}
+	void shm_lock(int offset, int n, int flags) override
+	{
+		_check_error(_parent->pMethods->xShmLock(_parent.get(), offset, n, flags));
+	}
+	void shm_barrier() noexcept override
+	{
+		_parent->pMethods->xShmBarrier(_parent.get());
+	}
+	void shm_unmap(int delete_flag) override
+	{
+		_check_error(_parent->pMethods->xShmUnmap(_parent.get(), delete_flag));
+	}
+	void fetch(sqlite3_int64 offset, int amount, void** buffer) override
+	{
+		_check_error(_parent->pMethods->xFetch(_parent.get(), offset, amount, buffer));
+	}
+	void unfetch(sqlite3_int64 offset, void* buffer) override
+	{
+		_check_error(_parent->pMethods->xUnfetch(_parent.get(), offset, buffer));
 	}
 
 private: