+++ ysqlite3/vfs/sqlite3_vfs_wrapper.hpp
@@ -40,7 +40,7 @@ public:
 		    reinterpret_cast<sqlite3_file*>(new std::uint8_t[_parent->szOsFile]{}),
 		    [](sqlite3_file* x) { delete[] reinterpret_cast<std::uint8_t*>(x); });
 		_check_error(_parent->xOpen(_parent, name, tmp_file.get(), flags, &output_flags));
-		return std::unique_ptr<file>{ new File{ format, std::move(tmp_file) } };
+		return std::unique_ptr<file>{ new File{ name, format, std::move(tmp_file) } };
 	}
 	void delete_file(const char* name, bool sync_directory) override
 	{